优化ui

+++ app/src/main/java/net/cd1369/tbs/android/data/entity/FavoriteEntity.kt
@@ -2,7 +2,7 @@ package net.cd1369.tbs.android.data.entity
 
 data class FavoriteEntity(
     val count: Int,
-    val createTime: Int,
+    val createTime: Long,
     val id: String,
     var list: List<ArticleEntity>? = mutableListOf<ArticleEntity>(),
     val name: String,
